pyproject.toml
[project]
name = "pointsmaxxer"
version = "0.1.0"
description = "Local flight points tracker - scrapes award availability, calculates CPP, and alerts on unicorn deals"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "playwright>=1.40.0",
    "pydantic>=2.5.0",
    "pyyaml>=6.0",
    "rich>=13.0.0",
    "typer>=0.9.0",
    "sqlalchemy>=2.0.0",
    "apscheduler>=3.10.0",
    "httpx>=0.25.0",
    "diskcache>=5.6.0",
    "tenacity>=8.2.0",
    "numpy>=1.21.0",
    "python-dateutil>=2.8.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.1.0",
]

[project.scripts]
pointsmaxxer = "pointsmaxxer.cli:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["pointsmaxxer"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
